Almost done.  Will need to complete AQUILONAQD-65 (disk should have a name parameter) and revamp the tests to finish.
The final issues were:
- A broken template-king on the broker (resolved indepently).
- A fun issue with struct.pack() in IPV4.py that stole 2 hours of
	my life.
- Direct access of the interface ip value (now a long) from
	show hostiplist.
- New fullnames for the hubs that are mixed case.
- A broken unit test for update_interface.

I suspect that commands without unit tests are probably still broken.

196/202 unit tests pass.

[git-p4: depot-paths = "//eai/aquilon/aqd/": change = 646111]

+++ 1.2.1/src/lib/python2.5/aquilon/aqdb/column_types/IPV4.py
@@ -47,7 +47,9 @@ class IPV4(types.TypeDecorator):
         return struct.unpack('L', inet_aton(dq))[0]
 
     def process_result_value(self, n, engine):
-        return inet_ntoa(struct.pack('L', n))
+        # Force the incoming Decimal to a long to prevent odd issues when
+        # struct.pack() tries it...
+        return inet_ntoa(struct.pack('L', long(n)))
 
     def copy(self):
         return IPV4(self.impl.length)
